:boom: :arrow_right: dependencies: Migrate to Pipenv
This commit migrates the existing Python project dependencies from the
various requirements files into a Pipenv file. I organized the packages
based on what I think is needed for a production installation of
Happiness Packets and the packages a developer would want while working.

I used a downstream Pipfile as a base:

    https://pagure.io/fedora-commops/fedora-happiness-packets/blob/master/f/Pipfile

Signed-off-by: Justin W. Flory <git@jwf.io>